+++ com/token/http/grant/code.js
@@ -93,7 +93,7 @@ exports = module.exports = function(ats, acs, logger, C) {
     });
 }
 
-exports['@implements'] = 'http://i.authnomicon.org/oauth2/token/http/GrantType';
+exports['@implements'] = 'module:oauth2orize.tokenRequestHandler';
 exports['@type'] = 'authorization_code';
 exports['@require'] = [
   'http://i.authnomicon.org/oauth2/AccessTokenService',