+++ test/token/http/grant/code.test.js
@@ -10,7 +10,7 @@ var oauth2orize = require('oauth2orize');
 describe('token/http/grant/code', function() {
   
   it('should be annotated', function() {
-    expect(factory['@implements']).to.equal('http://i.authnomicon.org/oauth2/token/http/GrantType');
+    expect(factory['@implements']).to.equal('module:oauth2orize.tokenRequestHandler');
     expect(factory['@type']).to.equal('authorization_code');
     expect(factory['@singleton']).to.be.undefined;
   });