+++ CHANGELOG.md
@@ -20,9 +20,11 @@ interface to `module:oauth2orize.RequestParametersProcessor`.
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseMode`
 interface to `module:oauth2orize.Responder`.
 - Renamed `http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters`
-to `module:oauth2orize.responseParametersFn`.
+interface to `module:oauth2orize.responseParametersFn`.
+- Renamed `http://i.authnomicon.org/oauth2/token/http/GrantType` interface to
+`module:oauth2orize.tokenRequestHandler`.
 - Renamed `http://i.authnomicon.org/oauth2/token/http/ResponseParameters`
-to `module:@authnomicon/oauth2.tokenResponseParametersFn`.
+interface to `module:@authnomicon/oauth2.tokenResponseParametersFn`.
 
 ## [0.0.4] - 2021-12-13
 ### Added